requirements.txt
easyocr
pandas
PyMuPDF
google-generativeai
certifi
customtkinter
pyinstaller
